re-implement FailedRequestException as HttpClientException
