+++ src/SourceClient.php
@@ -6,7 +6,7 @@ namespace SmartAssert\ApiClient;
 
 use SmartAssert\ApiClient\Data\Source\SourceInterface;
 use SmartAssert\ApiClient\Exception\Error\ErrorException;
-use SmartAssert\ApiClient\Exception\Http\FailedRequestException;
+use SmartAssert\ApiClient\Exception\Http\HttpClientException;
 use SmartAssert\ApiClient\Exception\Http\HttpException;
 use SmartAssert\ApiClient\Exception\Http\NotFoundException;
 use SmartAssert\ApiClient\Exception\Http\UnauthorizedException;
@@ -33,7 +33,7 @@ readonly class SourceClient
      *
      * @return SourceInterface[]
      *
-     * @throws FailedRequestException
+     * @throws HttpClientException
      * @throws HttpException
      * @throws IncompleteResponseDataException
      * @throws NotFoundException
@@ -77,7 +77,7 @@ readonly class SourceClient
      * @param non-empty-string $apiKey
      * @param non-empty-string $id
      *
-     * @throws FailedRequestException
+     * @throws HttpClientException
      * @throws HttpException
      * @throws IncompleteResponseDataException
      * @throws NotFoundException
@@ -95,7 +95,7 @@ readonly class SourceClient
      * @param non-empty-string $apiKey
      * @param non-empty-string $id
      *
-     * @throws FailedRequestException
+     * @throws HttpClientException
      * @throws HttpException
      * @throws IncompleteResponseDataException
      * @throws NotFoundException
@@ -115,7 +115,7 @@ readonly class SourceClient
      * @param non-empty-string $id
      *
      * @throws ErrorException
-     * @throws FailedRequestException
+     * @throws HttpClientException
      * @throws HttpException
      * @throws IncompleteResponseDataException
      * @throws NotFoundException
